Added functions to pyBach
I added appendStartTime and appendEndTime to pyBach module. These will allows the users to specify the start and end times, in posix time, for parsing PushShift. I also added the getRedditSubLists function. I need to decided if I want to write a wrapper function or not. Need to time how long it takes to sample all submissions in a subreddit.